Fixed permission issue with oradata
A recent change has introduced volumes to docker-oracle-xe-11g that broke the initialization of the oracle instance.

This is due to a defect with the recursive copy of the original oradata: cp -r does not retain the original attributes of the files such as owner and group. When oracle initializes, it doesn't have permission to access the oradata. Instead of cp -r, cp -a does the same as cp -r but preserves the attributes. This effectively retains the permission settings of the original oradata.